make all cpp tests dynamic linked to libpaddle.so
