+++ paddle/fluid/framework/new_executor/CMakeLists.txt
@@ -27,7 +27,8 @@ if(WITH_GPU
     COMMAND wget -nc --no-check-certificate
             https://paddle-ci.gz.bcebos.com/new_exec/lm_main_program
     COMMAND wget -nc --no-check-certificate
-            https://paddle-ci.gz.bcebos.com/new_exec/lm_startup_program)
+            https://paddle-ci.gz.bcebos.com/new_exec/lm_startup_program
+    WORKING_DIRECTORY "${CC_TESTS_DIR}")
 
   # all operators used in the program
   set(OPS
@@ -58,16 +59,11 @@ if(WITH_GPU
   # All deps of the operators above, part of GLOB_OPERATOR_DEPS.
   set(OP_DEPS generator softmax selected_rows_functor jit_kernel_helper
               concat_and_split cross_entropy)
+  cc_test(standalone_executor_test SRCS standalone_executor_test.cc)
 
-  cc_test(
-    standalone_executor_test
-    SRCS standalone_executor_test.cc
-    DEPS standalone_executor operator op_registry executor ${OPS} ${OP_DEPS})
-  set_tests_properties(standalone_executor_test PROPERTIES TIMEOUT 100)
-
-  add_dependencies(standalone_executor_test download_program)
-  if(WITH_PROFILER)
-    target_link_libraries(standalone_executor_test profiler)
-    add_dependencies(standalone_executor_test profiler)
-  endif()
+  # add_dependencies(standalone_executor_test download_program)
+  # if(WITH_PROFILER)
+  #   target_link_libraries(standalone_executor_test profiler)
+  #   add_dependencies(standalone_executor_test profiler)
+  # endif()
 endif()
